refactor: move method to configure sslreloadtask in sslreload package
