+++ dropwizard-core/src/main/java/io/dropwizard/core/sslreload/SslReloadTask.java
@@ -1,13 +1,16 @@
 package io.dropwizard.core.sslreload;
 
+import io.dropwizard.jetty.MutableServletContextHandler;
 import io.dropwizard.jetty.SslReload;
 import io.dropwizard.servlets.tasks.Task;
-
+import io.dropwizard.core.setup.Environment;
 import java.io.PrintWriter;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.HashSet;
+import java.util.Set;
 
 /** A task that will refresh all ssl factories with up to date certificate information */
 public class SslReloadTask extends Task {
@@ -15,6 +18,18 @@ public class SslReloadTask extends Task {
 
     protected SslReloadTask() {
         super("reload-ssl");
+    }
+
+    // New method to encapsulate setting reloaders based on the environment
+    public void configureReloaders(Environment environment) {
+        final Set<SslReload> reloaders = new HashSet<>();
+        reloaders.addAll(getReloaders(environment.getApplicationContext()));
+        reloaders.addAll(getReloaders(environment.getAdminContext()));
+        this.reloader = Collections.unmodifiableSet(reloaders); // Immutable set for safety
+    }
+
+    private Collection<SslReload> getReloaders(MutableServletContextHandler handler) {
+        return handler.getServer().getBeans(SslReload.class);
     }
 
     @Override
